# docs/fs/cas_app_api.openapi.yaml
...
openapi: "3.1.0"
info:
  title: CAS Application API
  version: "2.4"
  description: >-
    HTTP API for the CAS Application backend. Provides workflow tools
    (directory scan, batch ingest with two-phase edge inference) that
    wrap SAGE Core operations behind UI-oriented entry points, mounted at
    the /app prefix. The server that hosts the SAGE Core API serves them,
    walking its own filesystem, under every profile, authenticated as the
    Core API is. A hosted deployment also serves them from the separate
    backend-for-frontend its single-page app is reached through; that
    process shares no filesystem with SAGE, so there the directory scan and
    path-based ingest answer 501 and file content is uploaded to the SAGE
    Core API's batch-ingest endpoint instead.
  contact:
    name: CAS Project
  license:
    name: Proprietary
servers:
  - url: https://cas.{baseDomain}
    description: >-
      A deployed CAS Application edge. Each deployment owns its own domain and
      Entra tenant, so `baseDomain` selects the target deployment.
    variables:
      baseDomain:
        default: cor.org
        description: Owned base domain of the target deployment.
  - url: http://localhost:8000
    description: Local profile (FastAPI).

# No `securitySchemes` block, and no operation-level `security`. A hosted
# deployment's backend-for-frontend requires the `cas_session` cookie opened by
...
# session with `auth_required` 401 before its parameters are bound or refused, so
...

...
paths:
...
  /app/scan:
    post:
      operationId: list_directory
      summary: Scan a filesystem directory and return file status with parsed metadata.
      description: >-
        Walks a directory (optionally bounded by max_depth), classifies each
        file by file extension against the vault's registered source
        adapters, parses filename metadata using the vault's configured
        filename_extraction rules, hashes file contents with SHA-256, and
        cross-references against the vault's graph store to determine
        per-file ingestion status.


        Status values returned per file:

        - **new**: hash is unknown to the vault; eligible for ingestion.

        - **modified**: source_path is known but content hash differs.

        - **unchanged**: source_path and content hash both match an
          existing document.

        - **no_adapter**: no registered adapter handles this extension.


        Used by the Ingest view to preview a directory before triggering
        batch ingestion.
      tags: [Scan]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "200":
          description: Scan results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResponse"
        "400":
          description: >-
            `invalid_directory`: `directory` does not exist or is not
            readable.

            `invalid_vault_id`: a vault id in the request is not
            well-formed.

            `unknown_parameter`: the request carries a body field or query
            parameter the operation does not declare.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: >-
            `auth_required`: a hosted deployment's backend-for-frontend
            received the request without a signed-in session. Sign in and retry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: >-
            `vault_not_found`: no vault registered with that id;
            `detail.available_vaults` lists the registered vaults.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "501":
          description: >-
            `local_profile_only`: a hosted deployment's backend-for-frontend
            received the request.
            Its backend shares no filesystem with SAGE, so it cannot walk a
            directory.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: >-
            `auth_not_configured`: a hosted deployment's backend-for-frontend
            received the request, and interactive sign-in is not configured for
            it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Batch ingest with SSE streaming
  # ---------------------------------------------------------------------------
  /app/ingest:
    post:
      operationId: bulk_ingest_document
      summary: Ingest a batch of files with two-phase edge inference, streamed via Server-Sent Events.
      description: >-
        Three-phase pipeline:


        - Phase 1: build the edge plan from the supplied parsed_metadata
          and the existing vault graph (version_chain Tier 1 supersedes,
          filename_code_match Tier 2 covers).

        - Phase 2: per-file sequential ingestion with metadata application
          and per-file error isolation (one failure does not abort the
          batch).

        - Phase 3: post-ingest edge resolution -- Tier 1 via link(),
          Tier 2 via staging_edges insertion. Planned edges can be dropped
          rather than created: edges referencing failed-ingest files, and
          Tier-1 supersessions refused because the target cannot be read,
          does not exist, or is in a lifecycle state that does not permit
          the transition. Drops are counted in the summary's
          `edges_dropped` and reported per edge in `edge_warnings`
          (EdgeWarning entries).


        Response is `text/event-stream` with two event types:


        - `progress` events emitted on file start, completion, and error,
          carrying file_index, total_files, filename, stage, status, and
          (on success) document_id or (on error) error message.

        - `summary` event emitted once at the end, carrying the
          batch summary (counts of new docs, edges created/staged/
          dropped, metadata pending, abstracts generated/deferred, per-file
          errors as BatchIngestFileError entries -- a typed SAGE error
          carrying its code and detail -- and edge warnings when any were
          produced).


        A failure of the batch itself, which can only surface after the 200
        is committed, ends the stream: it emits no further events -- there is
        no `error` event variant -- and closes the connection without a
        `summary`. A stream that ends without one did not complete, and the
        cause is logged server-side rather than sent. Clients should treat
        the `summary` event, not the end of the stream, as the completion
        signal.


        See ScanResponse for how files are first classified before being
        passed to this endpoint.
      tags: [Ingest]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestRequest"
      responses:
        "200":
          description: Server-Sent Event stream of progress and summary events.
          content:
            text/event-stream:
              schema:
                type: string
                description: >-
                  SSE stream. Each event is a `data:` line with a JSON
                  payload conforming to ProgressEvent or SummaryEvent.
                  See components.schemas.ProgressEvent and
                  components.schemas.SummaryEvent.
        "400":
          description: >-
            `empty_file_list`: `files` was empty. Choose at least one file
            or skip the call.

            `invalid_vault_id`: a vault id in the request is not
            well-formed.

            `invalid_document_date`: a document date in the request is not
            a well-formed calendar date (`YYYY-MM-DD`).

            `unknown_parameter`: the request carries a body field or query
            parameter the operation does not declare.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: >-
            `auth_required`: a hosted deployment's backend-for-frontend
            received the request without a signed-in session. Sign in and retry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: >-
            `vault_not_found`: no vault registered with that id;
            `detail.available_vaults` lists the registered vaults.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "501":
          description: >-
            `local_profile_only`: a hosted deployment's backend-for-frontend
            received the request.
            Its backend shares no filesystem with SAGE, so it cannot read files
            by path; upload their content to the SAGE Core API's batch-ingest
            endpoint instead.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: >-
            `auth_not_configured`: a hosted deployment's backend-for-frontend
            received the request, and interactive sign-in is not configured for
            it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Interactive sign-in (backend-for-frontend)
  #
  # The OIDC redirect callback (GET /app/auth/callback) is a browser-facing
  # redirect mechanism, not a documented JSON API operation, so it is excluded
  # from this surface (include_in_schema=False on the route; allowlisted in the
  # OpenAPI conformance gate's infra-paths).
  # ---------------------------------------------------------------------------
...
components:
  schemas:
...
    ScanRequest:
      type: object
      additionalProperties: false
      required: [vault_id, directory]
      description: Request body for the directory-scan endpoint.
      properties:
        vault_id:
          type: string
          description: Target vault identifier.
        directory:
          type: string
          description: >-
            Absolute or user-relative directory path. Surrounding single
            or double quotes are stripped to tolerate paste-with-quotes
            from terminals or finder.
        max_depth:
          type: [integer, "null"]
          minimum: 0
          description: >-
            Optional recursion ceiling. 0 = scan the directory root only,
            1 = root plus one subdirectory level, etc. Null = the server's
            default depth ceiling; a scan cut by that default is reported
            as truncated.
    ScanResponse:
      type: object
      required: [files, warnings]
      description: Response body for the directory-scan endpoint; lists per-file scan results and any non-fatal warnings.
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/ScanResultResponse"
          description: Per-file scan results in directory-walk order.
        warnings:
          type: array
          items:
            type: string
          description: >-
            Non-fatal scan warnings (e.g. unreadable file, malformed
            symlink). Files that produced warnings may also appear in
            files[] with status "no_adapter" or be omitted depending on
            the warning class.
        truncated:
          type: boolean
          description: >-
            True when a scan ceiling (file count, hashed bytes, or the
            default depth) cut the walk short; the warnings list names
            the ceiling that fired.
    ScanResultResponse:
      type: object
      required: [file_path, file_hash, source_modified_at, parsed_metadata, sage_status]
      description: One per-file result inside a scan response; carries identity, parsed metadata, and the vault-relative ingestion status.
      properties:
        file_path:
          type: string
          description: Absolute file path on disk.
        file_hash:
          type: string
          description: SHA-256 hex digest of file contents.
        source_modified_at:
          type: string
          format: date-time
          description: File mtime (st_mtime) as an ISO 8601 timestamp.
        source_type:
          type: [string, "null"]
          description: >-
            Source artifact format matching this extension, or null
            when no registered adapter handles the extension.
        parsed_metadata:
          $ref: "#/components/schemas/ParsedMetadata"
        sage_status:
          type: string
          enum: [new, modified, unchanged, no_adapter]
          description: Vault-relative status for this file. See operation description.
    ParsedMetadata:
      type: object
      additionalProperties: false
      required: [title]
      description: Filename-derived metadata for one file; surfaced inside scan results and used as caller-authoritative input to ingest.
      properties:
        title:
          type: string
          description: Human-readable title extracted from the filename or content.
        date:
          type: [string, "null"]
          format: date
          description: Document calendar date (YYYY-MM-DD) extracted from filename.
        project:
          type: [string, "null"]
          description: >-
            Project identifier extracted from filename when the vault's
            project_identifier rule matches a segment.
        codes:
          type: array
          items:
            type: string
          description: >-
            Code identifiers extracted from filename via known_code_patterns
            (e.g. "PV07", "CD-2"). Order preserved from the filename.
        version:
          type: [string, "null"]
          description: Canonical version label (e.g. "v1.2.0", "v3.0").
        doc_type:
          type: [string, "null"]
          description: >-
            Document type resolved by the vault's keyword_to_doc_type and
            code_to_doc_type rules. Null if no rule matches.

    # -------------------------------------------------------------------------
    # Ingest
    # -------------------------------------------------------------------------
    IngestRequest:
      type: object
      additionalProperties: false
      required: [vault_id, files]
      description: Request body for the batch-ingest endpoint; identifies the target vault and the files to ingest with optional edge inference.
      properties:
        vault_id:
          type: string
          description: Target vault identifier.
        files:
          type: array
          items:
            $ref: "#/components/schemas/IngestFileItem"
          description: Files to ingest. Empty list returns 400.
        infer_edges:
          type: boolean
          default: true
          description: >-
            When true, the post-ingest phase runs version_chain (Tier 1
            supersedes) and filename_code_match (Tier 2 covers) edge
            inference. When false, edges are not inferred; ingestion is
            otherwise unchanged.
        dry_run:
          type: boolean
          default: false
          description: >-
            When true, evaluate every file and report one preview or
            one refusal per entry without persisting anything. No
            source is retained, no projection, indexing or abstraction
            runs, no record is written, and edge inference does not
            run whatever `infer_edges` says. Items are evaluated
            against committed state as it stood at batch start, so a
            sequential dependency within one batch is not reflected:
            two entries carrying identical bytes each report no
            duplicate.
    IngestFileItem:
      type: object
      additionalProperties: false
      required: [file_path, source_type]
      description: One file entry in a batch-ingest request; pairs a path with its source_type and optional caller-supplied parsed metadata.
      properties:
        file_path:
          type: string
          description: Absolute file path on disk.
        source_type:
          type: string
          description: Source artifact format name to use for this file.
        parsed_metadata:
          anyOf:
            - $ref: "#/components/schemas/ParsedMetadata"
            - type: "null"
          description: >-
            Optional caller-supplied parsed metadata for this file; when
            present, used as caller-authoritative input to ingest.
...
    ErrorResponse:
      type: object
      required: [code, message]
      description: >-
        Common error envelope returned by every CAS Application API endpoint
        for non-2xx responses. Mirrors the SAGE Core API ErrorResponse and
        reuses the same Pydantic class (sage.models.schemas.ErrorResponse).
      properties:
        code:
          type: string
          description: Machine-readable error code (e.g., "invalid_lifecycle_transition", "document_not_found", "editor_permission_denied").
        message:
          type: string
          description: Human-readable error description.
        detail:
          type: [object, "null"]
          description: Additional context. Structure varies by error type (e.g., current_state and attempted_action for lifecycle transition errors).
        read_meta:
          type: object
          description: >-
            Self-describing read markers (CAS-ADR-039). `success=False` on
            the error envelope, parallel to the success-response marker.

    # -------------------------------------------------------------------------
    # Interactive sign-in (backend-for-frontend)
    # -------------------------------------------------------------------------
...